requirements.txt
sqlite3
bcrypt
pycrypto